flask/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
langchain-groq==0.1.0
langchain==0.1.0
langchain-core==0.1.0
langchain-community==0.1.0
faiss-cpu==1.7.2
langchain-google-genai==0.0.3